moving layout index tests to index_test.go and adding test coverage to the index Push method
